Merge in some of the material from the testing bit.
I've not yet added in the rename function because I'm having second thoughts.
Still all very sketchy.

I'm standardising on a period naming scheme (rather than underscores) because
under pressure that's still what I reach for, and because it matches
most base functions better.  This is potentially a huge problem for R though;
there is so much variation even within the base packages.